add InstanceRefinement test
