+++ org.eclipse.winery.model.adaptation/pom.xml
@@ -14,7 +14,7 @@
         <checkstyle.config.location>../checkstyle.xml</checkstyle.config.location>
         <main.basedir>${project.parent.basedir}</main.basedir>
     </properties>
-    
+
     <dependencies>
         <dependency>
             <groupId>org.eclipse.winery</groupId>
@@ -46,13 +46,6 @@
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-params</artifactId>
             <version>${junit.jupiter.version}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.sshd</groupId>
-            <artifactId>sshd-sftp</artifactId>
-            <version>2.6.0</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
