+++ org.eclipse.winery.repository/pom.xml
@@ -269,6 +269,11 @@
             <artifactId>commons-validator</artifactId>
             <version>1.6</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.sshd</groupId>
+            <artifactId>sshd-core</artifactId>
+            <version>2.6.0</version>
+        </dependency>
     </dependencies>
     <build>
         <resources>
